Introduce workaround for label breakpoint issue
In gcc 8 and 9 I observe that setting a breakpoint on the break_here
labels is unreliable, and sometimes results in a transfer of control
to the code right before the label, instead of right after. This workaround
causes gdb to run to directly after the subsequent dummy_function,
where the test function's locals are still in scope.